Improve the code Refactoring in various classes.
