+++ app/src/main/java/me/ccrama/redditslide/Activities/AlbumPager.java
@@ -125,7 +125,7 @@ public class AlbumPager extends FullScreenActivity
         if (id == R.id.comments) {
             int adapterPosition = getIntent().getIntExtra(MediaView.ADAPTER_POSITION, -1);
             finish();
-            SubmissionsView.datachanged(adapterPosition);
+            SubmissionsView.dataChanged(adapterPosition);
             //getIntent().getStringExtra(MediaView.SUBMISSION_SUBREDDIT));
             //SubmissionAdapter.setOpen(this, getIntent().getStringExtra(MediaView.SUBMISSION_URL));
         }
@@ -614,7 +614,7 @@ public class AlbumPager extends FullScreenActivity
                         @Override
                         public void onClick(View v) {
                             getActivity().finish();
-                            SubmissionsView.datachanged(adapterPosition);
+                            SubmissionsView.dataChanged(adapterPosition);
                         }
                     });
                 } else {
