+++ app/src/main/java/me/ccrama/redditslide/Activities/RedditGallery.java
@@ -89,7 +89,7 @@ public class RedditGallery extends FullScreenActivity implements FolderChooserDi
             mToolbar.findViewById(R.id.grid).callOnClick();
         }
         if (id == R.id.comments) {
-            SubmissionsView.datachanged(adapterPosition);
+            SubmissionsView.dataChanged(adapterPosition);
             finish();
         }
         if (id == R.id.external) {
