+++ app/src/main/java/me/ccrama/redditslide/Activities/RedditGalleryPager.java
@@ -98,7 +98,7 @@ public class RedditGalleryPager extends FullScreenActivity
         if (id == R.id.comments) {
             int adapterPosition = getIntent().getIntExtra(MediaView.ADAPTER_POSITION, -1);
             finish();
-            SubmissionsView.datachanged(adapterPosition);
+            SubmissionsView.dataChanged(adapterPosition);
             //getIntent().getStringExtra(MediaView.SUBMISSION_SUBREDDIT));
             //SubmissionAdapter.setOpen(this, getIntent().getStringExtra(MediaView.SUBMISSION_URL));
         }
@@ -388,7 +388,7 @@ public class RedditGalleryPager extends FullScreenActivity
                     @Override
                     public void onClick(View v) {
                         getActivity().finish();
-                        SubmissionsView.datachanged(adapterPosition);
+                        SubmissionsView.dataChanged(adapterPosition);
                     }
                 });
             } else {
